Made base Block class
