+++ game/js/objects/girderMarker.js
@@ -1,6 +1,6 @@
 var game = window.game;
 
-var Girder = require( "./girder" );
+var Girder = require( "./blocks" ).Girder;
 
 var COLLISION_GROUPS = require( "../consts/collisionGroups" );
 var EPSILON = require( "../consts" ).EPSILON;
